requirements.txt
gunicorn
Django==3.2.19
setuptools<=65.5.1
django-cors-headers>=3.7.0
dpd-static-support>=0.0.5
pandas>=1.0
whitenoise>=5.0.1
docutils>=0.16
python-markdown-math>=0.8
pints @ git+https://github.com/pints-team/pints@i1403-multi-outputs#egg=pints
myokit @ git+https://github.com/MichaelClerx/myokit@main#egg=myokit
django-extensions>=3.1.1
jsonfield>=3.1.0
djangorestframework>=3.12.4
djoser>=2.1.0
python-memcached>=1.16.0
psycopg2-binary>=2.9.1
dj-database-url>=0.5.0
django-polymorphic>=3.1.0
drf-writable-nested>=0.6.3
celery>=5.1.2
tdigest>=0.5.2.2
pymc3>=3.11.5
django-auth-ldap>=4.1.0
pyyaml>=6.0.0
drf-spectacular>=0.26.2
openpyxl>=3.1.2